UPBGE: int instead of short to store glsl flag settings
https://github.com/UPBGE/upbge/issues/1511

+++ source/gameengine/Ketsji/KX_KetsjiEngine.h
@@ -85,7 +85,7 @@ enum class KX_DebugOption
 };
 
 typedef struct {
-	short glslflag;
+	int glslflag;
 } GlobalSettings;
 
 /**
